Splitting ServerHandler to separate file
