+++ .editorconfig
@@ -1,0 +1,11 @@
+root = true
+[*]
+indent_style=tab
+indent_size=tab
+tab_width=2
+end_of_line=lf
+charset=utf-8
+trim_trailing_whitespace=true
+max_line_length=120
+insert_final_newline=true
+
