+++ Cargo.toml
@@ -3,7 +3,7 @@ description = "Rust http server using JSONRPC 2.0."
 homepage = "https://github.com/debris/jsonrpc-http-server"
 license = "MIT"
 name = "jsonrpc-http-server"
-version = "5.1.0"
+version = "6.1.0"
 authors = ["debris <marek.kotewicz@gmail.com>"]
 keywords = ["jsonrpc", "json-rpc", "json", "rpc", "server"]
